+++ packages/medusa-js/src/resources/carts.ts
@@ -1,10 +1,12 @@
 import {
   StoreCartsRes,
+  StoreOrdersRes,
   StorePostCartReq,
   StorePostCartsCartPaymentSessionReq,
   StorePostCartsCartPaymentSessionUpdateReq,
   StorePostCartsCartReq,
   StorePostCartsCartShippingMethodReq,
+  StoreSwapsRes,
 } from "@medusajs/medusa"
 import { ResponsePromise } from "../typings"
 import BaseResource from "./base"
@@ -34,9 +36,11 @@ class CartsResource extends BaseResource {
    * The completion of a cart can be performed idempotently with a provided header Idempotency-Key.
    * If not provuided, we will generate one for the request.
    * @param {string} cart_id is required
-   * @return {ResponsePromise<StoreCartsRes>}
+   * @return {ResponsePromise<StoreCartsRes | StoreSwapsRes | StoreOrdersRes>}
    */
-  complete(cart_id: string): ResponsePromise<StoreCartsRes> {
+  complete(
+    cart_id: string
+  ): ResponsePromise<StoreCartsRes | StoreSwapsRes | StoreOrdersRes> {
     const path = `/store/carts/${cart_id}/complete`
     return this.client.request("POST", path)
   }
@@ -148,14 +152,16 @@ class CartsResource extends BaseResource {
   /**
    * Updates the payment method
    * @param {string} cart_id is required
+   * @param {string} provider_id is required
    * @param {StorePostCartsCartPaymentSessionUpdateReq} payload is required
    * @return {ResponsePromise<StoreCartsRes>}
    */
   updatePaymentSession(
     cart_id: string,
+    provider_id: string,
     payload: StorePostCartsCartPaymentSessionUpdateReq
   ): ResponsePromise<StoreCartsRes> {
-    const path = `/store/carts/${cart_id}/payment-session/update`
+    const path = `/store/carts/${cart_id}/payment-sessions/${provider_id}`
     return this.client.request("POST", path, payload)
   }
 }