+++ packages/medusa-source-shopify/src/services/__mocks__/product-variant.js
@@ -1,0 +1,17 @@
+export const ProductVariantServiceMock = {
+  withTransaction: function () {
+    return this
+  },
+  create: jest.fn().mockImplementation((_data) => {
+    return Promise.resolve({})
+  }),
+  update: jest.fn().mockImplementation((_data) => {
+    return Promise.resolve()
+  }),
+  delete: jest.fn().mockImplementation((_data) => {
+    return Promise.resolve()
+  }),
+  addOption: jest.fn().mockImplementation((id, title) => {
+    return Promise.resolve()
+  }),
+}