+++ packages/medusa-source-shopify/src/services/__mocks__/shopify-product.js
@@ -1,0 +1,10 @@
+import { medusaProducts } from "./test-products"
+
+export const ShopifyProductServiceMock = {
+  withTransaction: function () {
+    return this
+  },
+  create: jest.fn().mockImplementation((_data) => {
+    return Promise.resolve(medusaProducts.ipod)
+  }),
+}