+++ packages/medusa/src/repositories/invite.ts
@@ -1,0 +1,5 @@
+import { EntityRepository, Repository } from "typeorm"
+import { Invite } from "../models/invite"
+
+@EntityRepository(Invite)
+export class InviteRepository extends Repository<Invite> {}