+++ www/reference/src/components/topbar.js
@@ -1,12 +1,12 @@
-import { Flex, Box, Link, Select } from "@theme-ui/components"
-import React, { useContext } from "react"
+import { Box, Flex, Link, Select } from "@theme-ui/components"
 import { navigate } from "gatsby-link"
-
+import React, { useContext } from "react"
 import GitHub from "../components/icons/github"
 import NavigationContext from "../context/navigation-context"
 import { convertToKebabCase } from "../utils/convert-to-kebab-case"
 import ChevronDown from "./icons/chevron-down"
 import Search from "./search"
+
 
 const Topbar = ({ data, api }) => {
   const { goTo, reset, currentSection } = useContext(NavigationContext)